--- src/main/java/com/driveu/server/domain/resource/domain/IconType.java
@@ -1,5 +1,0 @@
-package com.driveu.server.domain.resource.domain;
-
-public enum IconType {
-    YOUTUBE, GITHUB
-}
