+++ src/main/java/com/driveu/server/domain/resource/domain/type/FileExtension.java
@@ -1,4 +1,4 @@
-package com.driveu.server.domain.resource.domain;
+package com.driveu.server.domain.resource.domain.type;
 
 public enum FileExtension {
     PDF, MD //Todo: 나중에 추가
