+++ src/main/java/com/driveu/server/domain/resource/domain/type/IconType.java
@@ -1,0 +1,5 @@
+package com.driveu.server.domain.resource.domain.type;
+
+public enum IconType {
+    YOUTUBE, GITHUB
+}
